+++ config/middleware/metrics_middleware.go
@@ -1,0 +1,27 @@
+package middleware
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"github.com/momentohq/client-sdk-go/config/logger"
+)
+
+type MetricsMiddleware struct {
+	Log          logger.MomentoLogger
+	requestCount map[string]uint64
+}
+
+func (mw *MetricsMiddleware) OnRequest(theRequest interface{}, _ context.Context) {
+	// Log request
+	if mw.requestCount == nil {
+		mw.requestCount = make(map[string]uint64)
+	}
+	mw.requestCount[fmt.Sprintf("%T", theRequest)]++
+	jsonStr, _ := json.MarshalIndent(mw.requestCount, "", "  ")
+	mw.Log.Info(fmt.Sprintf("%+v\n", string(jsonStr)))
+}
+
+func (mw *MetricsMiddleware) OnResponse(_ map[string]string) {
+	return
+}
